Format with prettier-plugin-solidity

+++ contracts/src/tokens/MockYearnVault.sol
@@ -3,26 +3,26 @@ pragma solidity ^0.8.4;
 import "../interfaces/yearnVault.sol";
 
 contract MockYearnVault is yVault {
-    constructor(address strategy) {}
+	constructor(address strategy) {}
 
-    function deposit(uint256 _amount) public override {}
+	function deposit(uint256 _amount) public override {}
 
-    function withdraw(uint256 amount) public override {}
+	function withdraw(uint256 amount) public override {}
 
-    function getPricePerFullShare() external pure override returns (uint) {
-        return 1;
-    }
-    function balanceOf(address who) external pure override returns (uint){
-        return 1;
-    }
+	function getPricePerFullShare() external pure override returns (uint256) {
+		return 1;
+	}
+
+	function balanceOf(address who) external pure override returns (uint256) {
+		return 1;
+	}
 }
 
+// function deposit(uint _amount) external;
 
-    // function deposit(uint _amount) external;
+// //returns # of tokens per share
+// function getPricePerFullShare() external view returns (uint);
 
-    // //returns # of tokens per share
-    // function getPricePerFullShare() external view returns (uint);
+// function withdraw(uint _shares) external;
 
-    // function withdraw(uint _shares) external;
-
-    // function balanceOf(address who) external view returns (uint);
+// function balanceOf(address who) external view returns (uint);
